src/totaltrackie/ui/icons.py
#  Copyright 2024 blsam
#
#  Permission is hereby granted, free of charge, to any person obtaining a copy of
#  this software and associated documentation files (the “Software”), to deal in
#  the Software without restriction, including without limitation the rights to use,
#  copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the
#  Software, and to permit persons to whom the Software is furnished to do so,
#  subject to the following conditions:
#
#  The above copyright notice and this permission notice shall be included in all
#  copies or substantial portions of the Software.
#
#  THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND,
#  EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
#  OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
#  NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
#  HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
#  WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
#  FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
#  OTHER DEALINGS IN THE SOFTWARE.
#

from enum import Enum, auto
from pathlib import Path

from PySide6.QtGui import QIcon


class IconResource(Enum):
    APP = auto()
    START = auto()
    ADD = auto()
    REMOVE = auto()
    EDIT = auto()
    OK = auto()
    CANCEL = auto()
    RELAX = auto()
    SETTINGS = auto()
    REPORT = auto()
    TEMPLATES = auto()

    def get_icon(self) -> QIcon:
        icon = QIcon(str(Path(__file__).absolute().parent.parent / "icons" / f"{self.name.lower()}.svg"))
        return icon
